requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
google-genai
httpx
python-multipart
PyPDF2
